requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.0
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.124.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
immutabledict==4.2.2
jiter==0.12.0
json_repair==0.54.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.3
langchain-core==1.1.2
langchain-openai==1.1.1
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
numpy==2.3.5
openai==2.9.0
openpyxl==3.1.5
OpenRoute==0.0.1
orjson==3.11.5
ormsgpack==1.12.0
ortools==9.14.6206
packaging==25.0
pandas==2.3.3
protobuf==6.31.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uuid_utils==0.12.0
uvicorn==0.38.0
xlrd==2.0.2
xxhash==3.6.0
zstandard==0.25.0
